# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Delivery Tracker API
  version: "1.0.0"
  description: |
    Multi-tenant API для трекинга доставок. Dispatch Manager'ы управляют
    брендами, грузами и поездками через `/admin/*`. Конечные получатели
    просматривают Share Page через `/share/:hash` (публично, без аутентификации).

    ## Аутентификация
    Сессия хранится в `httpOnly` cookie `auth` (JWT HS256, TTL 7 дней).
    Установить через `POST /auth/login`.

    ## Multi-tenancy
    Каждый Brand принадлежит ровно одному Dispatch Manager'у.
    Brand-scoped эндпоинты `/admin/b/:brandSlug/*` возвращают 404,
    если slug не принадлежит текущему пользователю.
servers:
  - url: http://localhost:3000
    description: Локальная разработка
  - url: https://admin.bluestarautotransport.com
    description: Продакшн (admin domain)
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth
      description: JWT токен, устанавливается через POST /auth/login
  schemas:
...
    Brand:
      type: object
      required: [id, slug, name, shareDomain]
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
          pattern: "^[a-z0-9](?:[a-z0-9-]*[a-z0-9])?$"
          minLength: 1
          maxLength: 64
          example: "blue-star-auto"
        name:
          type: string
          example: "Blue Star Auto Transport"
        shareDomain:
          type: string
          example: "tracking.bluestarautotransport.com"
    BrandCreate:
      type: object
      required: [slug, name, shareDomain]
      properties:
        slug:
          type: string
          pattern: "^[a-z0-9](?:[a-z0-9-]*[a-z0-9])?$"
          minLength: 1
          maxLength: 64
        name:
          type: string
          minLength: 1
          maxLength: 120
        shareDomain:
          type: string
          description: Валидный hostname в нижнем регистре. Должен быть уникальным.
          example: "tracking.bluestarautotransport.com"
    BrandDnsStatus:
      type: object
      required: [resolved, expected, actual]
      properties:
        resolved:
          type: boolean
          description: true если shareDomain резолвится в тот же IP что и admin-домен
        expected:
          type: array
          items:
            type: string
          description: IP-адреса admin-домена
          example: ["193.23.201.57"]
        actual:
          type: array
          items:
            type: string
          description: IP-адреса shareDomain
          example: ["193.23.201.57"]
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "not found"
...
  responses:
    Unauthenticated:
      description: Cookie `auth` отсутствует или JWT невалиден
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "unauthenticated"
    NotFound:
      description: Ресурс не найден (или не принадлежит текущему пользователю)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "not found"
    NoContent:
      description: Успешно, без тела ответа
...
paths:
...
  /admin/brands:
    get:
      tags: [brands]
      summary: Список своих брендов
      operationId: listBrands
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Бренды текущего Dispatch Manager'а
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Brand"
        "401":
          $ref: "#/components/responses/Unauthenticated"

    post:
      tags: [brands]
      summary: Создать бренд
      operationId: createBrand
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandCreate"
      responses:
        "201":
          description: Бренд создан
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Brand"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "409":
          description: Slug или share domain уже используется
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "slug or share domain already in use"
  /admin/brands/{slug}:
    delete:
      tags: [brands]
      summary: Удалить бренд
      operationId: deleteBrand
      security:
        - cookieAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/brands/{slug}/dns-status:
    get:
      tags: [brands]
      summary: Статус DNS-резолвинга share domain
      operationId: getBrandDnsStatus
      description: |
        Сравнивает A-записи `shareDomain` с A-записями admin-домена.
        `resolved: true` означает, что CNAME настроен правильно.
      security:
        - cookieAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: DNS статус
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandDnsStatus"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
  /admin/geocode:
    get:
      tags: [brands]
      summary: Геокодинг адреса (Mapbox proxy)
      operationId: geocode
      description: |
        Проксирует запрос к Mapbox Geocoding v5. Mapbox токен остаётся на сервере.
        Запрос требует минимум 2 символа в `q`.
      security:
        - cookieAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
          example: "New York"
      responses:
        "200":
          description: Список найденных мест
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  required: [label, lng, lat]
                  properties:
                    label:
                      type: string
                      example: "New York, New York, United States"
                    lng:
                      type: number
                      example: -74.006
                    lat:
                      type: number
                      example: 40.7128
        "401":
          $ref: "#/components/responses/Unauthenticated"

  # ── Brand-scoped: Dashboard ──────────────────────────────────────────────
...